'Not' filters should return true for nulls. Fix filter chip message
