+++ projects/table-builder/src/lib/components/gen-filter-displayer/gen-filter-displayer.component.html
@@ -18,7 +18,7 @@
 
   <mat-chip-list>
     <mat-chip *ngFor="let filter of filters$ | async " (dblclick)="currentFilters.push(filter)"  (removed)="tableState.removeFilter(filter.filterId)">
-      {{ filter.key | keyDisplay | async  }} {{filter.filterType}} {{ filter.filterValue | formatValue: filter.key : filter.filterType  | async }}
+      {{ filter.key | keyDisplay | async  }} {{filter.filterType | formatFilterType : filter.filterValue}} {{ filter.filterValue | formatFilterValue: filter.key : filter.filterType  | async }}
       <mat-icon matChipRemove>cancel</mat-icon>
     </mat-chip>
   </mat-chip-list>
