--- projects/table-builder/src/lib/pipes/format-value.ts
@@ -1,39 +1,0 @@
-import { Pipe, PipeTransform } from '@angular/core';
-import { TableStore } from '../classes/table-store';
-import { Observable } from 'rxjs';
-import { map } from 'rxjs/operators';
-import { FieldType } from '../interfaces/report-def';
-import { DatePipe } from '@angular/common';
-import { FilterType } from '../enums/filterTypes';
-import { FilterInfo } from '../classes/filter-info';
-
-
-@Pipe({name: 'formatValue'})
-export class FormatValuePipe implements PipeTransform {
-  constructor( public tableState: TableStore , private datePipe: DatePipe) {
-  }
-  transform(value: any, key: string, filterType: FilterType): Observable<string> {
-    return this.tableState.getMetaData$(key).pipe(
-      map( md => {
-        if(filterType === FilterType.Or) {
-          const filters = value as FilterInfo [];
-          const res =  filters.reduce( (prev,curr) => prev  + curr.filterValue + ' or ' ,'');
-          return res.substr(0, res.length - 4);
-        }
-        if(value && (filterType === FilterType.NumberIn || filterType === FilterType.StringIn)){
-          return (value as {value:any}[]).map(v => v.value).join(', ') ?? value;
-        }
-        if(filterType === FilterType.NumberBetween){
-          if(md.fieldType === FieldType.Date){
-            value = this.datePipe.transform(value.Start, 'MM/dd/yy')+ ' - ' + this.datePipe.transform(value.End, 'MM/dd/yy');
-          }else{
-            value = value.Start + ' - ' + value.End;
-          }
-        } else if(md.fieldType === FieldType.Date){
-          value = this.datePipe.transform(value, 'MM/dd/yy');
-        }
-        return value
-      })
-    );
-  }
-}
